Moved table-export
