+++ ui/packages/catalog-ui-search/src/main/webapp/react-component/table-export/index.tsx
@@ -12,4 +12,4 @@
  * <http://www.gnu.org/licenses/lgpl.html>.
  *
  **/
-export { default } from './table-export'
+export { default } from './container'
